+++ chelsea.gemspec
@@ -37,10 +37,11 @@ Gem::Specification.new do |spec|
 
 
   spec.add_dependency "tty-font", "~> 0.5.0"
+  spec.add_dependency "tty-spinner", "~> 0.9.3"
   spec.add_dependency "slop", "~> 4.8.0"
   spec.add_dependency "pastel", "~> 0.7.2"
   spec.add_dependency "rest-client", "~> 2.0.2"
-  # spec.add_dependency "semantic", "~> 1.6.1"
+  spec.add_dependency "semantic", "~> 1.6.1"
 
   # spec.add_development_dependency "bundler", "~> 2.0"
   spec.add_development_dependency "rake", "~> 10.0"
