+++ spec/unit/gems_spec.rb
@@ -1,14 +1,14 @@
-# require 'chelsea/commands/gems'
+require 'chelsea/gems'
 
-# RSpec.describe Chelsea::Commands::Gems do
-#   it "executes `gems` command successfully" do
-#     output = StringIO.new
-#     file = nil
-#     options = {}
-#     command = Chelsea::Commands::Gems.new(file, options)
+RSpec.describe Chelsea::Gems do
+  it "executes `gems` command successfully" do
+    output = StringIO.new
+    file = nil
+    options = {}
+    command = Chelsea::Gems.new(file, options)
 
-#     command.execute(output: output)
+    command.execute(output: output)
 
-#     expect(output.string).to eq("OK\n")
-#   end
-# end
+    expect(output.string).to eq("OK\n")
+  end
+end
